requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-docx==0.8.11
openpyxl==3.1.2
reportlab==4.0.4
pefile==2023.2.7
pdfminer.six==20221105
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.2
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2 